+++ services/vpn/v1alpha1api/configuration.go
@@ -21,15 +21,12 @@ func NewConfiguration() *config.Configuration {
 		Debug:         false,
 		Servers: config.ServerConfigurations{
 			{
-				URL:         "https://vpn.api.{region}stackit.cloud",
+				URL:         "https://vpn.api.stackit.cloud",
 				Description: "No description provided",
 				Variables: map[string]config.ServerVariable{
 					"region": {
 						Description:  "No description provided",
-						DefaultValue: "eu01.",
-						EnumValues: []string{
-							"eu01.",
-						},
+						DefaultValue: "global",
 					},
 				},
 			},