+++ services/kms/CHANGELOG.md
@@ -1,5 +1,8 @@
-## v1.10.0
+## v1.11.0
 - **Feature:** Introduce enums for various attributes
+
+# v1.10.0
+- `v1betaapi`: This API was removed after the announced deprecation period.
 
 ## v1.9.0
 - **Improvement:** Use new WaitHandler struct for all wait handlers