+++ services/logme/CHANGELOG.md
@@ -1,5 +1,9 @@
-## v0.29.0
+## v0.30.0
 - **Feature:** Introduce enums for various attributes
+
+## v0.29.0
+- `v1api`
+  - **Improvement**: Use new `WaiterHandler` struct in the LogMe WaitHandler
 
 ## v0.28.2
 - **Dependencies:** Bump STACKIT SDK core module from `v0.25.0` to `v0.26.0`