+++ services/observability/CHANGELOG.md
@@ -1,5 +1,8 @@
-## v0.22.0
+## v0.23.0
 - **Feature:** Introduce enums for various attributes
+
+## v0.22.0
+- **Breaking change:** Updated `InstanceSensitiveData` model because of misbehaving API
 
 ## v0.21.2
 - **Dependencies:** Bump STACKIT SDK core module from `v0.25.0` to `v0.26.0`